+++ ts/subject/ValueSubject.ts
@@ -1,23 +1,32 @@
 import { Subject } from './Subject.class'
-import { OperatorFunction, SubjectSubscriber, Subscription } from './subject.utils'
+import { SubjectSubscriber, Subscription } from './subject.utils'
 
 type ValueSubjectSubscriber<T> = (
   value: T,
   subscription: Subscription<T>,
-  // ...value: T[]
-  // subscription: Subscription,
 ) => unknown
 
 export class ValueSubject<T> extends Subject<T> {
-  constructor(public value: T) {
+  declare _value: T
+
+  constructor(value: T) {
     super(value)
+  }
+
+  get value() {
+    return this._value
+  }
+
+  set value(newValue) {
+    this._value = newValue;
+    this.set(newValue)
   }
 
   subscribe(callback: ValueSubjectSubscriber<T>) {
     const subscription = super.subscribe(callback as SubjectSubscriber<T>)
     
     // Call the callback immediately with the current value
-    callback(this.value, subscription)
+    callback(this._value, subscription)
 
     return subscription
   }