+++ ts/tag/Tag.class.ts
@@ -1,7 +1,7 @@
 import { Counts } from '../interpolations/interpolateTemplate'
 import { State } from '../state'
 import { InterpolatedTemplates } from '../interpolations/interpolations'
-import { InterpolateSubject } from './update/processFirstSubject.utils'
+import { InterpolateSubject, ValueTypes } from './update/processFirstSubject.utils'
 import { TemplaterResult } from '../TemplaterResult.class'
 import { TagValues } from './html'
 
@@ -27,7 +27,7 @@ export interface TagTemplate {
 }
 
 export class Tag {
-  isTagClass = true
+  tagJsType = ValueTypes.tag
 
   // present only when an array. Populated by Tag.key()
   memory = {