+++ ts/tag/cloneValueArray.function.ts
@@ -1,23 +1,23 @@
 import { Tag } from './Tag.class'
 import { deepClone } from '../deepFunctions'
-import { isTagArray, isTagClass, isTagComponent, isTagTemplater } from '../isInstance'
 import { TemplaterResult } from '../TemplaterResult.class'
+import { ValueTypes, getValueType } from './update/processFirstSubject.utils'
 
 export function cloneValueArray<T>(values: (T | Tag | Tag[])[]): T[] {
   return values.map((value) => {
     const tag = value as Tag
 
-    if(isTagComponent(value)) {
-      const tagComponent = value as TemplaterResult
-      return deepClone(tagComponent.props)
-    }
+    switch(getValueType(value)) {
+      case ValueTypes.tagComponent:
+        const tagComponent = value as TemplaterResult
+        return deepClone(tagComponent.props)
+      
+      case ValueTypes.tag:
+      case ValueTypes.templater:
+        return cloneValueArray(tag.values)
 
-    if(isTagClass(tag) || isTagTemplater(tag)) {
-      return cloneValueArray(tag.values)
-    }
-
-    if(isTagArray(tag)) {
-      return cloneValueArray(tag as unknown as Tag[])
+      case ValueTypes.tagArray:
+        return cloneValueArray(tag as unknown as Tag[])
     }
 
     return deepClone(value)